fix test cases
