+++ internal/exec/terraform_clean_test.go
@@ -5,74 +5,8 @@ import (
 	"testing"
 
 	"github.com/cloudposse/atmos/pkg/schema"
-	"github.com/pkg/errors"
 	"github.com/stretchr/testify/require"
 )
-
-// test ExecuteTerraform clean command.
-func TestCLITerraformClean(t *testing.T) {
-	err := os.Unsetenv("ATMOS_CLI_CONFIG_PATH")
-	if err != nil {
-		t.Fatalf("Failed to unset 'ATMOS_CLI_CONFIG_PATH': %v", err)
-	}
-
-	err = os.Unsetenv("ATMOS_BASE_PATH")
-	if err != nil {
-		t.Fatalf("Failed to unset 'ATMOS_BASE_PATH': %v", err)
-	}
-
-	// Capture the starting working directory
-	startingDir, err := os.Getwd()
-	if err != nil {
-		t.Fatalf("Failed to get the current working directory: %v", err)
-	}
-
-	defer func() {
-		// Change back to the original working directory after the test
-		if err := os.Chdir(startingDir); err != nil {
-			t.Fatalf("Failed to change back to the starting directory: %v", err)
-		}
-	}()
-
-	// Define the work directory and change to it
-	workDir := "../../tests/fixtures/scenarios/terraform-sub-components"
-	if err := os.Chdir(workDir); err != nil {
-		t.Fatalf("Failed to change directory to %q: %v", workDir, err)
-	}
-
-	var infoApply schema.ConfigAndStacksInfo
-	infoApply.SubCommand = "apply"
-	infoApply.ComponentType = "terraform"
-	infoApply.Stack = "staging"
-	infoApply.Component = "component-1"
-	infoApply.ComponentFromArg = "component-1"
-	err = ExecuteTerraform(infoApply)
-	require.NoError(t, err)
-	infoApply.Component = "component-2"
-	infoApply.ComponentFromArg = "component-2"
-	err = ExecuteTerraform(infoApply)
-	require.NoError(t, err)
-	files := []string{
-		"../../components/terraform/mock-subcomponents/component-1/.terraform",
-		"../../components/terraform/mock-subcomponents/component-1/terraform.tfstate.d/staging-component-1/terraform.tfstate",
-		"../../components/terraform/mock-subcomponents/component-1/component-2/.terraform",
-		"../../components/terraform/mock-subcomponents/component-1/component-2/terraform.tfstate.d/staging-component-2/terraform.tfstate",
-	}
-	success, file := verifyFileExists(t, files)
-	if !success {
-		t.Fatalf("File %s does not exist", file)
-	}
-	var cleanInfo schema.ConfigAndStacksInfo
-	cleanInfo.SubCommand = "clean"
-	cleanInfo.ComponentType = "terraform"
-	cleanInfo.AdditionalArgsAndFlags = []string{"--force"}
-	err = ExecuteTerraform(cleanInfo)
-	require.NoError(t, err)
-	success, file = verifyFileDeleted(t, files)
-	if !success {
-		t.Fatalf("File %s should not exist", file)
-	}
-}
 
 func verifyFileDeleted(t *testing.T, files []string) (bool, string) {
 	for _, file := range files {
@@ -325,15 +259,4 @@ func TestCollectComponentsDirectoryObjects(t *testing.T) {
 			}
 		})
 	}
-}
-
-func verifyFileExists(t *testing.T, files []string) (bool, string) {
-	success := true
-	for _, file := range files {
-		if _, err := os.Stat(file); errors.Is(err, os.ErrNotExist) {
-			t.Errorf("Reason: Expected file does not exist: %q", file)
-			return false, file
-		}
-	}
-	return success, ""
 }
